requirements.txt
numpy==1.26.4
scikit-surprise==1.1.3
streamlit==1.35.0
pandas==2.0.3
joblib==1.4.2